Add support for selection property (beta) to the typings generator (#27)
Add support to the typings generator for the new `selection` property,
which is added as a beta feature in Mendix 9.23. Since the feature is
still in beta, we are omitting it from the changelog for now.

## Relevant changes

Depending on the configuration of the `selection` property, the typings
generator will now emit:
- A client prop of (union) type `SelectionSingleValue` and/or
`SelectionMultiValue`. Selection type `None` and/or a linked data source
with `required="false"` results in an optional prop.
- A preview prop of (union) type `"None"`, `"Single"` and/or `"Multi"`.
A linked data source with `required="false"` will cause `"None"` to be
included in the type.

Unit tests have been added to check various combinations of the above.