chore: give a better name to the prop
This property means if the column is available to the user at all. While name `visible`
resembles with conditional visibility it makes more harm in the context of internal
structures by appearing related to the show/hide functionality of the data grid itself.
Like is column hidden by the user or not. While this means that column is not possible to
interact with at all. Hence renaming it.
The underlying properties are not renamed as those are coming from XML.