fix: mobx in modern client
1. Patch mobx to isolate its global store. In modern client setup, calling
config() to isolate is not enough, this happens too late and gives an error.
Hence the fix is to use separate namespace for mobx globals (patch out version).

2. Patch mobx-react-lite to prevent usage of use-sync-external-store shim.
This shim is causing issues with modern client as it uses `require("react")`
which is not processed at this point and fails in runtime.